+++ src/splat/util/symbols.py
@@ -334,6 +334,8 @@ def initialize_spim_context(all_segments: "List[Segment]") -> None:
 
     from ..segtypes.common.code import CommonSegCode
 
+    global_segments_after_overlays: List[CommonSegCode] = []
+
     for segment in all_segments:
         if not isinstance(segment, CommonSegCode):
             # We only care about the VRAMs of code segments
@@ -363,6 +365,10 @@ def initialize_spim_context(all_segments: "List[Segment]") -> None:
                 global_vram_end = segment.vram_end
             elif global_vram_end < segment.vram_end:
                 global_vram_end = segment.vram_end
+
+                if len(overlay_segments) > 0:
+                    # Global segment *after* overlay segments?
+                    global_segments_after_overlays.append(segment)
 
             if global_vrom_start is None:
                 global_vrom_start = segment.rom_start
@@ -401,6 +407,7 @@ def initialize_spim_context(all_segments: "List[Segment]") -> None:
             global_vrom_start, global_vrom_end, global_vram_start, global_vram_end
         )
 
+        overlaps_found = False
         # Check the vram range of the global segment does not overlap with any overlay segment
         for ovl_segment in overlay_segments:
             assert (
@@ -411,9 +418,27 @@ def initialize_spim_context(all_segments: "List[Segment]") -> None:
                 and global_vram_end > ovl_segment.vramStart
             ):
                 log.write(
-                    f"Warning: the vram range ([0x{ovl_segment.vramStart:08X}, 0x{ovl_segment.vramEnd:08X}]) of the non-global segment at rom address 0x{ovl_segment.vromStart:X} overlaps with the global vram range ([0x{global_vram_start:08X}, 0x{global_vram_end:08X}])",
+                    f"Error: the vram range ([0x{ovl_segment.vramStart:08X}, 0x{ovl_segment.vramEnd:08X}]) of the non-global segment at rom address 0x{ovl_segment.vromStart:X} overlaps with the global vram range ([0x{global_vram_start:08X}, 0x{global_vram_end:08X}])",
                     status="warn",
                 )
+                overlaps_found = True
+        if overlaps_found:
+            log.write(
+                f"Many overlaps between non-global and global segments were found.",
+            )
+            log.write(
+                f"This is usually caused by missing `exclusive_ram_id` tags on segments that have a higher vram address than other `exclusive_ram_id`-tagged segments"
+            )
+            if len(global_segments_after_overlays) > 0:
+                log.write(
+                    f"These segments are the main suspects for missing a `exclusive_ram_id` tag:",
+                    status="warn",
+                )
+                for seg in global_segments_after_overlays:
+                    log.write(f"    '{seg.name}', rom: 0x{seg.rom_start:06X}")
+            else:
+                log.write(f"No suspected segments??", status="warn")
+            log.error("Stopping due to the above errors")
 
     # pass the global symbols to spimdisasm
     for segment in all_segments: